server/requirements.txt
fastapi
uvicorn[standard]
groq
razorpay
python-multipart
python-dotenv
pydantic
numpy
